Adds missing methods into the CMS API and adds default implementations in the ActiveMQ-CPP code for them that throw UnsupportedOperationExceptions.  Cleans up a bunch of code in preperation for the release, fixes warnings and removes debug code.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@775980 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/decaf/internal/nio/LongArrayBufferTest.cpp
@@ -511,7 +511,7 @@ void LongArrayBufferTest::testPutIndexed() {
     for( std::size_t i = 0; i < testBuffer1->capacity(); i++ ) {
         CPPUNIT_ASSERT( testBuffer1->position() == 0 );
         LongBuffer& ret = testBuffer1->put(i, i );
-        CPPUNIT_ASSERT( testBuffer1->get(i) == i );
+        CPPUNIT_ASSERT( testBuffer1->get(i) == (long long)i );
         CPPUNIT_ASSERT( &ret == testBuffer1 );
     }
